requirements.txt
PyPDF2
requests
python-dotenv